requirements.txt
streamlit>=1.25.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
youtube-transcript-api>=1.0.3
langchain>=0.3.24
langchain_openai>=0.3.14
langchain-google-genai>=0.6.15
streamlit_mermaid >= 0.3.0
notion-client>=2.3.0
streamlit-local-storage>=0.0.25
langchain-openai>=0.3.18
openai>=1.84.0
apify>=2.5.0
pytubefix>=9.1.1